DAO-226 Move useCountdown hook into its own file
