+++ src/components/timer/timer.tsx
@@ -1,10 +1,8 @@
-import { useState, useEffect, useRef } from 'react';
 import classNames from 'classnames';
-import { getDays, getHours, getMinutes, getSeconds } from '../../utils/generic';
+import { useCountdown } from './hooks';
 import globalStyles from '../../styles/global-styles.module.scss';
 import styles from './timer.module.scss';
 import { format } from 'date-fns';
-import { usePrevious } from '../../utils';
 
 interface Props {
   deadline: Date;
@@ -16,14 +14,6 @@ interface Props {
 export const DATE_FORMAT = 'do MMMM yyyy';
 
 export const formatDeadline = (deadline: Date) => format(deadline, DATE_FORMAT);
-
-interface Countdown {
-  days: string;
-  hours: string;
-  minutes: string;
-  seconds: string;
-  dateDiff: number;
-}
 
 const Timer = (props: Props) => {
   const { deadline, size = 'normal', showDeadline } = props;
@@ -72,51 +62,3 @@ const Timer = (props: Props) => {
 };
 
 export default Timer;
-
-export function useCountdown(deadline: Date, onDeadlineExceeded?: () => void) {
-  const [countdown, setCountdown] = useState<Countdown>(() => calculateCountdown(deadline));
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      const countdown = calculateCountdown(deadline);
-      setCountdown(countdown);
-
-      if (countdown.dateDiff === 0) {
-        clearInterval(timer);
-      }
-    }, 1000);
-
-    return () => {
-      clearInterval(timer);
-    };
-  }, [deadline]);
-
-  const callbackRef = useRef(onDeadlineExceeded);
-  useEffect(() => {
-    callbackRef.current = onDeadlineExceeded;
-  });
-
-  const { dateDiff } = countdown;
-  const previousDateDiff = usePrevious(dateDiff) || 0;
-  useEffect(() => {
-    // We trigger the callback when we have gone past the deadline (i.e. from a positive diff to a zero diff)
-    if (previousDateDiff > 0 && dateDiff === 0) {
-      callbackRef.current && callbackRef.current();
-    }
-  }, [previousDateDiff, dateDiff]);
-
-  return countdown;
-}
-
-function calculateCountdown(deadline: Date): Countdown {
-  const now = new Date().getTime();
-  const dateDiff = Math.max(deadline.getTime() - now, 0);
-
-  return {
-    days: getDays(dateDiff),
-    hours: getHours(dateDiff),
-    minutes: getMinutes(dateDiff),
-    seconds: getSeconds(dateDiff),
-    dateDiff,
-  };
-}
